requirements.txt
matplotlib
numpy
opencv-python-headless==4.5.4.60
pandas
Pillow
streamlit
